refactor: topology
